Add support for user secrets
