+++ api/README.md
@@ -10,3 +10,8 @@ Get OAS3 specification file
 https://localhost:44323/swagger/v1/swagger.json
 ```
 
+
+# User Secrets
+In development mode, sensitive credentials can be used without adding to a configuration file that may accidently added 
+to source control. See [Safe storage of app secrets in development in ASP.NET Core](https://docs.microsoft.com/en-us/aspnet/core/security/app-secrets?view=aspnetcore-3.1&tabs=windows)
+for more information.  User secrets are only used for local development. At present, no secrets are required.
