+++ api/src/BcGov.Malt.Web/Program.cs
@@ -1,4 +1,6 @@
+using System.Reflection;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 
 namespace BcGov.Malt.Web
@@ -20,11 +22,24 @@ namespace BcGov.Malt.Web
         /// <summary>
         /// Creates the host builder.
         /// </summary>
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureWebHostDefaults(webBuilder =>
+        public static IHostBuilder CreateHostBuilder(string[] args)
+        {
+            var builder = Host.CreateDefaultBuilder(args)
+                .ConfigureWebHostDefaults(webBuilder => { webBuilder.UseStartup<Startup>(); })
+                .ConfigureAppConfiguration((hostingContext, config) =>
                 {
-                    webBuilder.UseStartup<Startup>();
+                    // Added before AddUserSecrets to let user secrets override environment variables.
+                    config.AddEnvironmentVariables();
+
+                    var env = hostingContext.HostingEnvironment;
+                    if (env.IsDevelopment())
+                    {
+                        var appAssembly = Assembly.Load(new AssemblyName(env.ApplicationName));
+                        config.AddUserSecrets(appAssembly, optional: true);
+                    }
                 });
+
+                return builder;
+        }
     }
 }
